# docs/swagger.yaml
basePath: /
definitions:
  model.AuthModel:
    properties:
      pass:
        type: string
      user:
        type: string
    type: object
  model.AuthResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      profiles:
        items:
          $ref: '#/definitions/model.Profile'
        type: array
      refToken:
        type: string
    type: object
  model.Profile:
    properties:
      default:
        type: boolean
      name:
        type: string
    type: object
  model.Token:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
    type: object
host: localhost:3333
info:
  contact: {}
  title: Stocks Calculator API
  version: "1.0"
paths:
  /createprofile:
    post:
      consumes:
      - application/json
      parameters:
      - description: Set as default profile (use 't' to enable)
        in: query
        name: default
        type: string
      - description: Profile
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Profile'
      responses:
        "200":
          description: OK
        "400":
          description: Missing profile name
          schema:
            type: string
        "409":
          description: Profile already exists
          schema:
            type: string
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create profile
      tags:
      - profiles
  /createuser:
    post:
      consumes:
      - application/json
      description: Create user with username and password
      parameters:
      - description: Create user credentials
        in: body
        name: authmodel
        required: true
        schema:
          $ref: '#/definitions/model.AuthModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AuthResponse'
        "400":
          description: incorrect request body format
          schema:
            type: string
        "409":
          description: user exists
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: CreateUser
      tags:
      - auth
  /getprofiles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Profile'
            type: array
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get profiles
      tags:
      - profiles
  /login:
    post:
      consumes:
      - application/json
      description: Log in with username and password
      parameters:
      - description: Login credentials
        in: body
        name: authmodel
        required: true
        schema:
          $ref: '#/definitions/model.AuthModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AuthResponse'
        "400":
          description: incorrect request body format
          schema:
            type: string
        "401":
          description: wrong credentials
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Login
      tags:
      - auth
  /reftoken:
    post:
      consumes:
      - application/json
      description: Grant access token for valid refresh token
      parameters:
      - description: Refresh token
        in: body
        name: tokenmodel
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Token'
        "400":
          description: no refresh token provided
          schema:
            type: string
        "404":
          description: refresh token is invalid
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: RefreshToken
      tags:
      - auth
  /removeprofile:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Profile
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Profile'
      responses:
        "200":
          description: OK
        "400":
          description: Missing profile name
          schema:
            type: string
        "404":
          description: Profile not found
          schema:
            type: string
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Remove profile
      tags:
      - profiles
  /setdefaultprofile:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Profile
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Profile'
      responses:
        "200":
          description: OK
        "400":
          description: Missing profile name
          schema:
            type: string
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Set default profile
      tags:
      - profiles
  /test:
    get:
      description: Grant access token for valid refresh token
      responses: {}
      security:
      - BearerAuth: []
      summary: Test
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"